Continued with video, don#t know exactly what to write...

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate(2));
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        return PostResource::collection($posts);
     }
 
     /**
@@ -24,8 +27,8 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
          $data = $request->validated();
-
-        $data['author_id'] = 1;
+ 
+        $data['author_id'] = $request->user()->id;
 
         $post = Post::create($data);
 
@@ -39,7 +42,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return response()->json(new PostResource($post));
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden!');
+        return new PostResource($post);
     }
 
     /**
@@ -47,6 +51,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden!');
         $data = $request->validate([
             
         ]);
@@ -61,6 +66,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, 'Access Forbidden!');
         $post->delete();
         return response()->noContent();
     }
